fix(#1274): consolidate BOPAlgoRemoveFeaturesTests into Issue497DefeaturingTests
- Merge 3 tests from BOPAlgoRemoveFeaturesTests into Issue497DefeaturingTests
- Rename tests to use 'defeature' terminology (replaces deprecated 'removeFeatures')
- Delete BOPAlgoRemoveFeaturesTests.swift
- All 8 tests in Issue497DefeaturingTests pass
- Gate scripts pass
